[CDS-23425]: Added Support for account level sdk default backoff strategy
