+++ 400-rest/src/main/java/software/wings/service/impl/aws/manager/AwsHelperServiceManager.java
@@ -24,6 +24,22 @@ import software.wings.service.impl.aws.model.AwsResponse;
 import com.google.inject.Singleton;
 import lombok.extern.slf4j.Slf4j;
 
+import software.wings.sm.ExecutionContext;
+
+import static software.wings.service.impl.aws.model.AwsConstants.BASE_DELAY_ACCOUNT_VARIABLE;
+import static software.wings.service.impl.aws.model.AwsConstants.MAX_BACKOFF_ACCOUNT_VARIABLE;
+import static software.wings.service.impl.aws.model.AwsConstants.MAX_ERROR_RETRY_ACCOUNT_VARIABLE;
+import static software.wings.service.impl.aws.model.AwsConstants.NULL_STR;
+import static software.wings.service.impl.aws.model.AwsConstants.THROTTLED_BASE_DELAY_ACCOUNT_VARIABLE;
+
+import static org.apache.commons.lang3.StringUtils.isBlank;
+
+import software.wings.beans.AmazonClientSDKDefaultBackoffStrategy;
+
+import com.google.common.annotations.VisibleForTesting;
+
+import software.wings.beans.AwsConfig;
+
 @Singleton
 @Slf4j
 @OwnedBy(CDP)
@@ -33,10 +49,58 @@ public class AwsHelperServiceManager {
       throw new InvalidRequestException(((ErrorNotifyResponseData) notifyResponseData).getErrorMessage(), USER);
     } else if (notifyResponseData instanceof RemoteMethodReturnValueData) {
       throw new InvalidRequestException(
-          getMessage(((RemoteMethodReturnValueData) notifyResponseData).getException()), USER);
+              getMessage(((RemoteMethodReturnValueData) notifyResponseData).getException()), USER);
     } else if (!(notifyResponseData instanceof AwsResponse)) {
       throw new InvalidRequestException(
-          format("Unknown response from delegate: [%s]", notifyResponseData.getClass().getSimpleName()), USER);
+              format("Unknown response from delegate: [%s]", notifyResponseData.getClass().getSimpleName()), USER);
     }
+  }
+
+
+  @VisibleForTesting
+  public static void setAmazonClientSDKDefaultBackoffStrategyIfExists(ExecutionContext context, AwsConfig awsConfig) {
+    if (!validateSDKDefaultBackoffStrategyAccountVariables(context)) {
+      return;
+    }
+
+    AmazonClientSDKDefaultBackoffStrategy sdkDefaultBackoffStrategy =
+            AmazonClientSDKDefaultBackoffStrategy.builder()
+                    .baseDelayInMs(resolveAccountVariable(context, BASE_DELAY_ACCOUNT_VARIABLE))
+                    .throttledBaseDelayInMs(resolveAccountVariable(context, THROTTLED_BASE_DELAY_ACCOUNT_VARIABLE))
+                    .maxBackoffInMs(resolveAccountVariable(context, MAX_BACKOFF_ACCOUNT_VARIABLE))
+                    .maxErrorRetry(resolveAccountVariable(context, MAX_ERROR_RETRY_ACCOUNT_VARIABLE))
+                    .build();
+    awsConfig.setAmazonClientSDKDefaultBackoffStrategy(sdkDefaultBackoffStrategy);
+    log.info("Using Amazon SDK default backoff strategy with account level values: {}",
+            sdkDefaultBackoffStrategy);
+  }
+
+  private static boolean validateSDKDefaultBackoffStrategyAccountVariables(ExecutionContext context) {
+    if (isRenderedExpressionBlank(context, BASE_DELAY_ACCOUNT_VARIABLE)
+            || isRenderedExpressionBlank(context, THROTTLED_BASE_DELAY_ACCOUNT_VARIABLE)
+            || isRenderedExpressionBlank(context, MAX_BACKOFF_ACCOUNT_VARIABLE)
+            || isRenderedExpressionBlank(context, MAX_ERROR_RETRY_ACCOUNT_VARIABLE)) {
+      return false;
+    }
+
+    try {
+      resolveAccountVariable(context, BASE_DELAY_ACCOUNT_VARIABLE);
+      resolveAccountVariable(context, THROTTLED_BASE_DELAY_ACCOUNT_VARIABLE);
+      resolveAccountVariable(context, MAX_BACKOFF_ACCOUNT_VARIABLE);
+      resolveAccountVariable(context, MAX_ERROR_RETRY_ACCOUNT_VARIABLE);
+    } catch (Exception ex) {
+      log.error("Not valid account level backoff strategy variables, msg: {}", ex.getMessage());
+      return false;
+    }
+    return true;
+  }
+
+  private static boolean isRenderedExpressionBlank(ExecutionContext context, final String expression) {
+    String renderedExpression = context.renderExpression(expression);
+    return isBlank(renderedExpression) || NULL_STR.equals(renderedExpression);
+  }
+
+  private static int resolveAccountVariable(ExecutionContext context, final String expression) {
+    return Integer.parseInt(context.renderExpression(expression));
   }
 }
