+++ 400-rest/src/test/java/software/wings/sm/states/CloudFormationStateTest.java
@@ -553,40 +553,4 @@ public class CloudFormationStateTest extends WingsBaseTest {
     assertThat(cloudFormationDeleteStackState.validateFields().size()).isEqualTo(0);
   }
 
-  @Test
-  @Owner(developers = IVAN)
-  @Category(UnitTests.class)
-  public void testSetAmazonClientSDKDefaultBackoffStrategyIfExists() {
-    when(executionContext.renderExpression(BASE_DELAY_ACCOUNT_VARIABLE)).thenReturn("100");
-    when(executionContext.renderExpression(THROTTLED_BASE_DELAY_ACCOUNT_VARIABLE)).thenReturn("500");
-    when(executionContext.renderExpression(MAX_BACKOFF_ACCOUNT_VARIABLE)).thenReturn("20000");
-    when(executionContext.renderExpression(MAX_ERROR_RETRY_ACCOUNT_VARIABLE)).thenReturn("5");
-
-    AwsConfig awsConfig = AwsConfig.builder().build();
-    cloudFormationCreateStackState.setAmazonClientSDKDefaultBackoffStrategyIfExists(executionContext, awsConfig);
-
-    AmazonClientSDKDefaultBackoffStrategy amazonClientSDKDefaultBackoffStrategy =
-        awsConfig.getAmazonClientSDKDefaultBackoffStrategy();
-    assertThat(amazonClientSDKDefaultBackoffStrategy.getBaseDelayInMs()).isEqualTo(100);
-    assertThat(amazonClientSDKDefaultBackoffStrategy.getThrottledBaseDelayInMs()).isEqualTo(500);
-    assertThat(amazonClientSDKDefaultBackoffStrategy.getMaxBackoffInMs()).isEqualTo(20000);
-    assertThat(amazonClientSDKDefaultBackoffStrategy.getMaxErrorRetry()).isEqualTo(5);
-  }
-
-  @Test
-  @Owner(developers = IVAN)
-  @Category(UnitTests.class)
-  public void testSetAmazonClientSDKDefaultBackoffStrategyIfExistsWithInvalidValues() {
-    when(executionContext.renderExpression(BASE_DELAY_ACCOUNT_VARIABLE)).thenReturn("not_valid");
-    when(executionContext.renderExpression(THROTTLED_BASE_DELAY_ACCOUNT_VARIABLE)).thenReturn("not_valid");
-    when(executionContext.renderExpression(MAX_BACKOFF_ACCOUNT_VARIABLE)).thenReturn("not_valid");
-    when(executionContext.renderExpression(MAX_ERROR_RETRY_ACCOUNT_VARIABLE)).thenReturn("not_valid");
-
-    AwsConfig awsConfig = AwsConfig.builder().build();
-    cloudFormationCreateStackState.setAmazonClientSDKDefaultBackoffStrategyIfExists(executionContext, awsConfig);
-
-    AmazonClientSDKDefaultBackoffStrategy amazonClientSDKDefaultBackoffStrategy =
-        awsConfig.getAmazonClientSDKDefaultBackoffStrategy();
-    assertThat(amazonClientSDKDefaultBackoffStrategy).isNull();
-  }
 }
